feat(routing): add TestRoutingContext
Exposes a public interface to rendering the routing context outside of an xstate-tree root, for tests/storybook etc
