+++ src/index.ts
@@ -29,6 +29,7 @@ export {
   matchRoute,
   useIsRouteActive,
   useRouteArgsIfActive,
+  TestRoutingContext,
 } from "./routing";
 export { loggingMetaOptions } from "./useService";
 export { lazy } from "./lazy";
